Fix 1.8 issues
